+++ src/com/intellij/idea/plugin/hybris/toolwindow/HybrisToolWindowFactory.kt
@@ -42,9 +42,8 @@ class HybrisToolWindowFactory : ToolWindowFactory, DumbAware {
         ).forEach { toolWindow.contentManager.addContent(it) }
     }
 
-    override fun shouldBeAvailable(project: Project): Boolean {
-        return HybrisProjectSettingsComponent.getInstance(project).isHybrisProject()
-    }
+    override fun isApplicable(project: Project) = HybrisProjectSettingsComponent.getInstance(project).isHybrisProject()
+    override fun shouldBeAvailable(project: Project) = HybrisProjectSettingsComponent.getInstance(project).isHybrisProject()
 
     private fun createTSContent(toolWindow: ToolWindow, panel: TSView) = with(toolWindow.contentManager.factory.createContent(panel, TS_ID, true)) {
         Disposer.register(toolWindow.disposable, panel)