+++ src/com/intellij/idea/plugin/hybris/toolwindow/system/bean/components/BSMetaPropertiesTable.kt
@@ -69,7 +69,7 @@ class BSMetaPropertiesTable private constructor(myProject: Project) :
             ),
             createColumn(
                 name = COLUMN_TYPE,
-                valueProvider = { attr -> attr.type }
+                valueProvider = { attr -> attr.flattenType }
             )
         )